changed fitness value calculation to 1/ticks

+++ src/swarm_sim/learning/GA.java
@@ -12,7 +12,7 @@ public class GA {
     public int PrevDirectionIndex = 5;
 
     public IChromosome currentChromosome;
-    public int currentFitness;
+    public double currentFitness;
 
     static GA instance;
